Ornstein-Uhlenbeck probability density plot done!

+++ simulation.c
@@ -13,7 +13,7 @@ double simulate_next_V(double last_v, double delta_t, double gamma, double D,
     return next_v;
 }
 
-double **simulate_V_values(double D, double gamma, int N, int N_t,
+double **simulate_V_values(double D, double gamma, long N, int N_t,
                            double delta_t, gsl_rng *r) {
     double v_max;
